+++ src/pages/Product/Product.jsx
@@ -10,7 +10,7 @@ return (
   <div className="section-card">
     <div className="container-card-head">
       <Link className="path" to="/Homepage">Home</Link>
-      <small> <i class="fas fa-angle-double-right"></i> </small>
+      <small> <i className="fas fa-angle-double-right"></i> </small>
       <Link className="path" to="/Product">Dog Treats</Link>
       <h2>Dog Treats</h2>
     </div>
